Add null protection to synonym return
The synonym code has recently been improved so that unrecognised keys return null, where previously they returned undefined. However, there is no protection for calls to this.synonyms in src/js/osweb/items/generic_response.js so when these unrecognised keys are returned null, the attempt to perform an array lookup on this.synonyms fails (because this.synonyms is null) and the experiment crashes.